refactor(editor): slash menu model 분리
